# registration_system_go/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Registration System Go API
  version: 1.0.0
  description: 赛事报名与球队管理 Go 后端 HTTP API。
servers:
  - url: http://127.0.0.1:18081
    description: 本地开发服务
...
paths:
...
  /api/v1/app/teams/{id}:
    get:
      operationId: getAppTeam
      tags: [App Teams]
      summary: 获取球队详情
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 球队详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppTeamDetailEnvelope"
        <<: *standardErrors
    patch:
      operationId: updateAppTeamProfile
      tags: [App Teams]
      summary: 更新球队资料（队长/领队）
      description: 仅该队 active 的队长或领队可操作。name 不传则保持不变，传入时必须非空且不超过 120 字符；description/logo_url 传空串或 null 视为清除。球队 status（冻结/激活）不可通过此接口修改。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                description: { type: string, nullable: true }
                logo_url: { type: string, nullable: true }
      responses:
        "200":
          description: 更新成功（空 data）
        <<: *standardErrors
    delete:
      operationId: deleteAppTeam
      tags: [App Teams]
      summary: 解散球队（仅队长）
      description: 仅该队 active 的队长本人可操作。软删除——teams.status 置为 dissolved，保留历史比赛/申请/支付数据。球队仍有未结束比赛（主/客队）或进行中约队申请时返回 409，可先调 GET /api/v1/app/teams/{id}/dissolve-blockers 查看阻塞项。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 解散成功（空 data）
        <<: *standardErrors
  /api/v1/app/teams/{id}/dissolve-blockers:
    get:
      operationId: getAppTeamDissolveBlockers
      tags: [App Teams]
      summary: 查询阻止球队解散的进行中引用（仅队长）
      description: 仅该队 active 的队长本人可查。返回未结束比赛（本队作为主/客队，is_host 标记能否在比赛详情页收尾）与进行中的约队申请（pending/selected，可在接约页撤回）；两个列表都为空时可以解散。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 阻塞引用列表（空列表表示可以解散）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppTeamDissolveBlockersEnvelope"
        <<: *standardErrors
  /api/v1/app/teams/{id}/join-password:
    put:
      operationId: updateAppTeamJoinPassword
      tags: [App Teams]
      summary: 更新入队密码（队长/领队）
      description: 仅该队 active 的队长或领队可操作。join_password 非空=设置/替换；空串=清除（开放加入）。语义与建队一致：仅用 trim 判空，密码保留原始值以 bcrypt 哈希存储，不回显。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [join_password]
              properties:
                join_password: { type: string }
      responses:
        "200":
          description: 更新成功（空 data）
        <<: *standardErrors
  /api/v1/app/teams/{id}/logo:
    post:
      operationId: uploadAppTeamLogo
      tags: [App Teams]
      summary: 上传球队 Logo（队长/领队）
      description: 仅该队 active 的队长或领队可操作。multipart 字段 file，≤1MB，jpg/png/webp，保存后返回新的 logo_url。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamLogoUploadEnvelope"
        <<: *standardErrors
  /api/v1/app/teams/{id}/members:
    get:
      operationId: listAppTeamMembers
      tags: [App Teams]
      summary: 获取球队成员
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 球队成员列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppTeamMemberListEnvelope"
        <<: *standardErrors
    post:
      operationId: addAppTeamMember
      tags: [App Teams]
      summary: 添加球队成员（队长/领队）
      description: 仅该队 active 的队长或领队可操作。role 仅支持 leader/vice_captain/member（缺省为 member），不能直接指定队长；目标用户必须存在且为 active。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: integer, format: int64 }
                role: { type: string, enum: [leader, vice_captain, member] }
      responses:
        "200":
          description: 添加成功（空 data）
        <<: *standardErrors
  /api/v1/app/teams/{id}/members/{user_id}:
    patch:
      operationId: updateAppTeamMember
      tags: [App Teams]
      summary: 修改球队成员角色/状态（队长/领队）
      description: 仅该队 active 的队长或领队可操作。role/status 至少传一个；role 仅支持 leader/vice_captain/member；目标成员为队长本人时返回 409，请使用移交队长。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: user_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [leader, vice_captain, member] }
                status: { type: string, enum: [active, inactive] }
      responses:
        "200":
          description: 修改成功（空 data）
        <<: *standardErrors
    delete:
      operationId: removeAppTeamMember
      tags: [App Teams]
      summary: 移除球队成员（队长/领队）
      description: 仅该队 active 的队长或领队可操作。队长本人不可移除，返回 409。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: user_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 移除成功（空 data）
        <<: *standardErrors
  /api/v1/app/teams/{id}/members/{user_id}/paid-membership:
    put:
      operationId: updateAppTeamMemberPaidMembership
      tags: [App Teams]
      summary: 更新队员付费会员与队费账户（队长/领队）
      description: 仅该队 active 的队长或领队可操作。手动切换付费会员标记；余额只能通过充值（fund-credits）、消费扣费（fund-consumptions）与冲正（fund-reversals）动作改变，直接设置余额的参数会被显式拒绝。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: user_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaidMembershipUpdateRequest"
      responses:
        "200":
          description: 更新成功（空 data）
        <<: *standardErrors
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TeamID:
      name: id
      in: path
      required: true
      description: 球队 ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
    SuccessBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          enum: [0]
        message:
          type: string
          example: ok
...
    MatchStatus:
      type: string
      enum: [registering, ongoing, ended, cancelled]
...
    ApplicationStatus:
      type: string
      enum: [pending, selected, rejected, withdrawn]
    TeamStatus:
      type: string
      enum: [active, frozen, dissolved]
    MemberStatus:
      type: string
      enum: [active, inactive]
    MemberRole:
      type: string
      enum: [captain, leader, vice_captain, member]
...
    TeamLogoUploadEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [logo_url]
              properties:
                logo_url:
                  type: string
...
    PaidMembershipUpdateRequest:
      type: object
      description: 仅支持手动切换付费会员标记；余额与最近充值时间由充值/消费/冲正动作维护。携带 balance_cents、last_recharge_at 等已废弃字段会被显式拒绝（422），不会静默忽略。
      required: [is_paid_member]
      properties:
        is_paid_member:
          type: boolean
          description: 该队员在当前球队是否为已缴队费会员；实际充值（微信到账或人工登记）会自动置为 true。
...
    AppTeamDetail:
      type: object
      required: [id, name, description, logo_url, captain_id, status, my_role]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        captain_id:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: "#/components/schemas/TeamStatus"
        my_role:
          $ref: "#/components/schemas/MemberRole"
    AppTeamMember:
      type: object
      required: [user_id, nickname, avatar_url, real_name, role, status, joined_at, balance_cents, is_paid_member, last_recharge_at]
      properties:
        user_id:
          type: integer
          format: int64
        nickname:
          type: string
        avatar_url:
          type: string
          nullable: true
        real_name:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/MemberRole"
        status:
          $ref: "#/components/schemas/MemberStatus"
        joined_at:
          type: string
          format: date-time
        balance_cents:
          type: integer
          format: int64
        is_paid_member:
          type: boolean
        last_recharge_at:
          type: string
          format: date-time
          nullable: true
...
    AppTeamDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AppTeamDetail"
    AppTeamDissolveBlockers:
      type: object
      required: [matches, applications]
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/AppTeamDissolveBlockerMatch"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/AppTeamDissolveBlockerApplication"
    AppTeamDissolveBlockerMatch:
      type: object
      required: [id, name, status, is_host]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: "#/components/schemas/MatchStatus"
        is_host:
          type: boolean
          description: 本队是否为发起方（主队）；只有主队能在比赛详情页收尾/取消比赛。
    AppTeamDissolveBlockerApplication:
      type: object
      required: [id, match_id, match_name, status]
      properties:
        id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        match_name:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
    AppTeamDissolveBlockersEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AppTeamDissolveBlockers"
    AppTeamMemberListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/AppTeamMember"
...